Adds ProjectMembership class
ProjectMembership defines the relationship between a User and a Project.
This opens up opportunity for invites to a project.

I still need to implement ownership of a project, and also invititations
to a project.

I also wish to implement an automatic invitation if a project has a
source control org which a user is already authenticated with.